improve code structure
