+++ bin/flutterpi_tool.dart
@@ -1,4 +1,4 @@
-import 'package:flutterpi_tool/flutterpi_tool.dart' as executable;
+import 'package:flutterpi_tool/src/executable.dart' as executable;
 
 Future<void> main(List<String> arguments) async {
   await executable.main(arguments);
